% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcs_functions.R
\name{dcs_plot}
\alias{dcs_plot}
\title{Plot dual-cutoff classifications}
\usage{
dcs_plot(score, group, status, cutoffs, score_label = "Score")
}
\arguments{
\item{score}{Numeric vector of scores.}

\item{group}{Numeric vector containing the reference-group codes \code{0} and
\code{1}.}

\item{status}{Classification status for each observation, using \code{TP}, \code{TN},
\code{FP}, \code{FN}, and \verb{Gray zone}.}

\item{cutoffs}{A \code{dcs_cutoffs} object returned by \code{\link[=get_cutoffs]{get_cutoffs()}}.}

\item{score_label}{Label for the y-axis.}
}
\value{
A ggplot object.
}
\description{
Plot dual-cutoff classifications
}
\examples{
data(clinitest)
cutoffs <- get_cutoffs(
  clinitest$score, clinitest$groups, 0.90, 0.90,
  n_bootstrap = 20, seed = 91000, direction = "lower", progress = FALSE
)
classified <- classify_subjects(clinitest$score, clinitest$groups, cutoffs)
dcs_plot(
  clinitest$score, clinitest$groups, classified$data$status, cutoffs
)
}
