% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcs_functions.R
\name{print.dcs_cutoffs}
\alias{print.dcs_cutoffs}
\title{Print dual-cutoff results}
\usage{
\method{print}{dcs_cutoffs}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{dcs_cutoffs} object.}

\item{digits}{Number of decimal places to print.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print dual-cutoff results
}
